requirements.txt
ultralytics
deepface
opencv-python
fastapi
uvicorn
numpy
easyocr
pillow
openai-whisper
librosa
soundfile
neo4j
redis
sqlalchemy
psycopg2-binary
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
python-multipart
streamlit
plotly
google-generativeai
python-dotenv